+++ spec/content_block_tools/presenters/block_presenters/contact/email_address_presenter_spec.rb
@@ -11,7 +11,8 @@ RSpec.describe ContentBlockTools::Presenters::BlockPresenters::Contact::EmailAdd
 
   let(:email_address) do
     {
-      "title": "Some email address",
+      "title": "Email",
+      "label": "Email us",
       "email_address": "foo@example.com",
     }
   end
@@ -20,8 +21,24 @@ RSpec.describe ContentBlockTools::Presenters::BlockPresenters::Contact::EmailAdd
     presenter = described_class.new(email_address, content_block:)
 
     expect(presenter.render).to have_tag("p") do
-      with_tag("span", text: "Some email address")
-      with_tag("a", text: "foo@example.com", with: { href: "mailto:foo@example.com", class: "email" })
+      with_tag("a", text: "Email us", with: { href: "mailto:foo@example.com", class: "email" })
+    end
+  end
+
+  describe "when the label is not present" do
+    let(:email_address) do
+      {
+        "title": "Email",
+        "email_address": "foo@example.com",
+      }
+    end
+
+    it "uses the email address as the link text" do
+      presenter = described_class.new(email_address, content_block:)
+
+      expect(presenter.render).to have_tag("p") do
+        with_tag("a", text: "foo@example.com", with: { href: "mailto:foo@example.com", class: "email" })
+      end
     end
   end
 
@@ -41,7 +58,6 @@ RSpec.describe ContentBlockTools::Presenters::BlockPresenters::Contact::EmailAdd
       result = presenter.render
 
       expect(result).to have_tag("p") do
-        with_tag("span", text: "Some email address")
         with_tag("a", text: "foo@example.com", with: { href: "mailto:foo@example.com?subject=My email&body=Body text here", class: "email" })
       end
     end
@@ -50,7 +66,8 @@ RSpec.describe ContentBlockTools::Presenters::BlockPresenters::Contact::EmailAdd
   describe "when a description is present" do
     let(:email_address) do
       {
-        "title": "Some email address",
+        "title": "Email us",
+        "label": "Some email address",
         "email_address": "foo@example.com",
         "description": "Description text",
       }
@@ -67,8 +84,7 @@ RSpec.describe ContentBlockTools::Presenters::BlockPresenters::Contact::EmailAdd
 
       expect(result).to have_tag("div", with: { class: "email-url-number" }) do
         with_tag("p") do
-          with_tag("span", text: "Some email address")
-          with_tag("a", text: "foo@example.com", with: { href: "mailto:foo@example.com", class: "email" })
+          with_tag("a", text: "Some email address", with: { href: "mailto:foo@example.com", class: "email" })
         end
 
         with_tag("p", text: "Description text")