+++ spec/content_block_tools/presenters/contact_presenter_spec.rb
@@ -70,7 +70,6 @@ RSpec.describe ContentBlockTools::Presenters::ContactPresenter do
             with_tag("div", with: { class: "vcard contact-inner" }) do
               with_tag("p", text: "My Contact")
               with_tag("p") do
-                with_tag("span", text: "Some email address")
                 with_tag("a", text: "foo@example.com", with: { href: "mailto:foo@example.com" })
               end
             end
@@ -96,7 +95,6 @@ RSpec.describe ContentBlockTools::Presenters::ContactPresenter do
         with_tag("div", with: { class: "contact" }) do
           with_tag("div", with: { class: "email-url-number" }) do
             with_tag("p") do
-              with_tag("span", text: "Some email address")
               with_tag("a", text: "foo@example.com", with: { href: "mailto:foo@example.com" })
             end
           end
@@ -328,7 +326,6 @@ RSpec.describe ContentBlockTools::Presenters::ContactPresenter do
             with_tag("div", with: { class: "vcard contact-inner" }) do
               with_tag("p", text: "My Contact")
               with_tag("p") do
-                with_tag("span", text: "Some contact form")
                 with_tag("a", text: "http://example.com", with: { href: "http://example.com" })
               end
             end
@@ -354,7 +351,6 @@ RSpec.describe ContentBlockTools::Presenters::ContactPresenter do
         with_tag("div", with: { class: "contact" }) do
           with_tag("div", with: { class: "email-url-number" }) do
             with_tag("p") do
-              with_tag("span", text: "Some contact form")
               with_tag("a", text: "http://example.com", with: { href: "http://example.com" })
             end
           end